start

+++ tasks/tbb/sorokin_a_multiplication_sparse_matrices_double_ccs/src/ops_tbb.cpp
@@ -17,7 +17,7 @@ void MultiplyCCS(const std::vector<double> &a_values, const std::vector<int> &a_
                  const std::vector<int> &a_col_ptr, const std::vector<double> &b_values,
                  const std::vector<int> &b_row_indices, int k, const std::vector<int> &b_col_ptr,
                  std::vector<double> &c_values, std::vector<int> &c_row_indices, int n, std::vector<int> &c_col_ptr) {
-  if (a_values.size() > m * k || b_values.size() > k * n) {
+  if (static_cast<int>(a_values.size()) > m * k || static_cast<int>(b_values.size()) > k * n) {
     throw std::invalid_argument("Invalid val pointer size");
   }
